Pull request da versão final da aplicação

+++ admin.py
@@ -47,16 +47,17 @@ def configurar_sprints():
     while True:
         event_config_sprints, values_config_sprints = window_config_sprints.read()
 
-        if event_config_sprints == sg.WINDOW_CLOSED or event_config_sprints == "Salvar":
+        if event_config_sprints == "Salvar":
+            # Persiste a quantidade de sprints
+            qtd_de_sprints = int(values_config_sprints["-SPRINTS-"])
+            dbj.set_qtd_de_sprints(qtd_de_sprints)
+
+            sg.popup(f"Quantidade de Sprints configurada para: {qtd_de_sprints}")
+            break
+        elif event_config_sprints == sg.WINDOW_CLOSED:
             break
 
     window_config_sprints.close()
-
-    # Persiste a quantidade de sprints
-    qtd_de_sprints = int(values_config_sprints["-SPRINTS-"])
-    dbj.set_qtd_de_sprints(qtd_de_sprints)
-
-    sg.popup(f"Quantidade de Sprints configurada para: {qtd_de_sprints}")
 
 # Solicitar informações do novo usuário
 def adicionar_usuario():
@@ -64,8 +65,8 @@ def adicionar_usuario():
         [sg.Text("Nome:      "), sg.Input(key="-NOME-", border_width=2, size=(25, 1))],
         [sg.Text("Matrícula: "), sg.Input(key="-MATRICULA-", border_width=2, size=(17, 1))],
         [sg.Text("Turma:     "), sg.Input(key="-TURMA-", border_width=2, size=(10, 1))],
-        [sg.Text("Times:     "), sg.Input(key="-TIMES-", border_width=2, size=(10, 1))],
-        [sg.Button("Adicionar", size=(10, 1))]
+        [sg.Text("Time:     "), sg.Input(key="-TIME-", border_width=2, size=(10, 1))],
+        [sg.Button("Adicionar"), sg.Button("Cancelar")]
     ]
 
     window_admin = sg.Window("Adicionar Usuário", layout_admin)
@@ -73,39 +74,39 @@ def adicionar_usuario():
     while True:
         event_admin, values_admin = window_admin.read()
 
-        if event_admin == sg.WINDOW_CLOSED or event_admin == "Adicionar":
+        if event_admin == "Adicionar":
+            # Carregar usuários existentes do arquivo JSON
+            with open("data.json", "r") as arquivo:
+                data = json.load(arquivo)
+
+            # Criar usuário com as informações digitadas
+            novo_usuario = {
+                "nome": values_admin["-NOME-"],
+                "matricula": values_admin["-MATRICULA-"],
+                "turma": values_admin["-TURMA-"],
+                "time": values_admin["-TIME-"]
+            }
+
+            if dbj.ra_exist(novo_usuario['matricula']):
+                sg.Popup('Já existe esse número de matrícula cadastrado!')
+                continue
+
+            # Adiciona o novo usuário na variável data
+            data['usuarios'].append({'matricula': novo_usuario['matricula'],
+                                    'nome': novo_usuario['nome'],
+                                    'turma': novo_usuario['turma'],
+                                    'time': novo_usuario['time']})
+
+            # Salvar a lista de usuários atualizada no arquivo JSON
+            with open("data.json", "w") as arquivo:
+                json.dump(data, arquivo, indent= 4)
+
+            sg.popup("Usuário adicionado com sucesso!")
+            break
+        elif event_admin == sg.WINDOW_CLOSED or event_admin == "Cancelar":
             break
 
     window_admin.close()
-
-    # Carregar usuários existentes do arquivo JSON
-    with open("data.json", "r") as arquivo:
-        data = json.load(arquivo)
-
-    usuarios = data["usuarios"]
-
-    # Criar usuário com as informações digitadas
-    novo_usuario = {
-        "nome": values_admin["-NOME-"],
-        "matricula": values_admin["-MATRICULA-"],
-        "turma": values_admin["-TURMA-"],
-        "times": values_admin["-TIMES-"]
-    }
-
-    # Adicionar o novo usuário à lista existente
-    usuarios.append(novo_usuario)
-
-    # Adiciona o novo usuário na variável data
-    data['usuarios'].append({'matricula': novo_usuario['matricula'],
-                             'nome': novo_usuario['nome'],
-                             'turma': novo_usuario['turma'],
-                             'time': novo_usuario['times']})
-
-    # Salvar a lista de usuários atualizada no arquivo JSON
-    with open("data.json", "w") as arquivo:
-        json.dump(data, arquivo, indent= 4)
-
-    sg.popup("Usuário adicionado com sucesso!")
 
 def editar_usuario(nome_usuario):
     # Carregar usuários do arquivo JSON
@@ -121,7 +122,9 @@ def editar_usuario(nome_usuario):
             layout_admin = [
                 [sg.Text("Novo nome: "), sg.Input(key="-NOVO_NOME-", default_text=usuario["nome"])],
                 [sg.Text("Nova matrícula: "), sg.Input(key="-NOVA_MATRICULA-", default_text=usuario["matricula"])],
-                [sg.Button("Atualizar")]
+                [sg.Text("Nova turma: "), sg.Input(key="-NOVA_TURMA-", default_text=usuario["turma"])],
+                [sg.Text("Novo time: "), sg.Input(key="-NOVO_TIME-", default_text=usuario["time"])],
+                [sg.Button("Atualizar"), sg.Button("Cancelar")]
             ]
 
             window_admin = sg.Window("Editar Usuário", layout_admin)
@@ -129,20 +132,23 @@ def editar_usuario(nome_usuario):
             while True:
                 event_admin, values_admin = window_admin.read()
 
-                if event_admin == sg.WINDOW_CLOSED or event_admin == "Atualizar":
+                if event_admin == event_admin == "Atualizar":
+                    # Atualizar informações do usuário
+                    usuario["nome"] = values_admin["-NOVO_NOME-"]
+                    usuario["matricula"] = values_admin["-NOVA_MATRICULA-"]
+                    usuario["turma"] = values_admin["-NOVA_TURMA-"]
+                    usuario["time"] = values_admin["-NOVO_TIME-"]
+
+                    # Salvar usuários atualizados no arquivo JSON
+                    with open("data.json", "w") as arquivo:
+                        json.dump(data, arquivo, indent=4)
+
+                    sg.popup("Usuário atualizado com sucesso!")
+                    break
+                elif event_admin == sg.WINDOW_CLOSED or event_admin == "Cancelar":
                     break
 
             window_admin.close()
-
-            # Atualizar informações do usuário
-            usuario["nome"] = values_admin["-NOVO_NOME-"]
-            usuario["matricula"] = values_admin["-NOVA_MATRICULA-"]
-
-            # Salvar usuários atualizados no arquivo JSON
-            with open("data.json", "w") as arquivo:
-                json.dump(data, arquivo)
-
-            sg.popup("Usuário atualizado com sucesso!")
             return
 
     sg.popup("Usuário não encontrado!")
